requirements.txt
absl-py==2.1.0
accelerate==0.31.0
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
AutoRAG==0.2.8
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
bert-score==0.3.13
bleach==6.1.0
blinker==1.8.2
bokeh==3.4.2
boto3==1.34.137
botocore==1.34.137
build==1.2.1
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.3
click==8.1.7
cloudpickle==3.0.0
cohere==5.5.8
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cramjam==2.8.3
cycler==0.12.1
dask==2024.6.2
dask-expr==1.1.6
dataclasses-json==0.6.7
datasets==2.20.0
debugpy==1.8.2
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
evaluate==0.4.2
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastavro==1.9.4
fastparquet==2024.5.0
filelock==3.15.4
FlagEmbedding==1.2.10
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.31.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.64.1
guidance==0.1.15
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
ipywidgets_bokeh==1.6.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kiwipiepy==0.17.1
kiwipiepy_model==0.17.0
kiwisolver==1.4.5
kubernetes==30.1.0
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.10
langchain-openai==0.1.13
langchain-text-splitters==0.2.2
langsmith==0.1.83
linkify-it-py==2.0.3
llama-cloud==0.0.6
llama-index==0.10.51
llama-index-agent-openai==0.2.7
llama-index-cli==0.1.12
llama-index-core==0.10.51
llama-index-embeddings-huggingface==0.2.2
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.2.2
llama-index-legacy==0.9.48
llama-index-llms-huggingface==0.2.4
llama-index-llms-ollama==0.1.5
llama-index-llms-openai==0.1.24
llama-index-llms-openai-like==0.1.3
llama-index-multi-modal-llms-openai==0.1.6
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.25
llama-index-readers-llama-parse==0.1.4
llama-index-retrievers-bm25==0.1.5
llama-parse==0.4.5
llmlingua==0.2.2
locket==1.0.0
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
minijinja==2.0.1
mkl==2021.4.0
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest_asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.18.1
openai==1.35.7
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
ordered-set==4.1.0
orjson==3.10.5
overrides==7.7.0
packaging==24.1
pandas==2.2.2
panel==1.4.4
param==2.1.1
parameterized==0.9.0
parso==0.8.4
partd==1.4.2
pickleshare==0.7.5
pillow==10.4.0
pip==24.0
platformdirs==4.2.2
portalocker==2.10.0
posthog==3.5.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.0
pydantic_core==2.20.0
pydeck==0.9.1
pydot==2.0.0
pyformlang==1.0.10
Pygments==2.18.0
pyparsing==3.1.2
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
pysbd==0.3.4
python-dateutil==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pyviz_comms==3.0.2
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
ragas==0.1.9
rank-bm25==0.2.2
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rouge_score==0.1.2
rpds-py==0.18.1
rsa==4.9
s3transfer==0.10.2
sacrebleu==2.4.2
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.14.0
seaborn==0.13.2
sentence-transformers==3.0.1
sentencepiece==0.2.0
setuptools==70.1.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.2
starlette==0.37.2
streamlit==1.36.0
striprtf==0.0.26
swifter==1.4.0
sympy==1.12.1
tabulate==0.9.0
tbb==2021.13.0
tenacity==8.4.2
text-generation==0.7.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tokenlog==0.0.2
toml==0.10.2
toolz==0.12.1
torch==2.3.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.3
typer==0.12.3
types-requests==2.32.0.20240622
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.1
uc-micro-py==1.0.3
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
wheel==0.43.0
widgetsnbextension==4.0.11
wrapt==1.16.0
xxhash==3.4.1
xyzservices==2024.6.0
yarl==1.9.4
zipp==3.19.2
